Renaming structure namespace to Metadata
