+++ src/Sulu/Bundle/ContentBundle/Content/Structure/SeoStructureExtension.php
@@ -11,7 +11,6 @@
 namespace Sulu\Bundle\ContentBundle\Content\Structure;
 
 use PHPCR\NodeInterface;
-use Sulu\Component\Content\StructureExtension\StructureExtension;
 use Sulu\Component\Content\Extension\AbstractExtension;
 
 /**
