+++ src/Sulu/Bundle/ContentBundle/Resources/config/structure.xml
@@ -4,8 +4,8 @@
            xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
 
     <parameters>
-        <parameter key="sulu_content.structure.factory.class">Sulu\Component\Content\Structure\Factory\StructureFactory</parameter>
-        <parameter key="sulu_content.structure.loader.xml.class">Sulu\Component\Content\Structure\Loader\XmlLoader</parameter>
+        <parameter key="sulu_content.structure.factory.class">Sulu\Component\Content\Metadata\Factory\StructureMetadataFactory</parameter>
+        <parameter key="sulu_content.structure.loader.xml.class">Sulu\Component\Content\Metadata\Loader\XmlLoader</parameter>
     </parameters>
 
     <services>
