+++ src/Sulu/Bundle/CoreBundle/Command/UpgradeSmartContentOperatorCommand.php
@@ -17,8 +17,8 @@ use Sulu\Component\Content\Block\BlockPropertyInterface;
 use Sulu\Component\Content\Block\BlockPropertyWrapper;
 use Sulu\Component\Content\Mapper\ContentMapperInterface;
 use Sulu\Component\Content\Mapper\Translation\TranslatedProperty;
-use Sulu\Component\Content\Structure;
-use Sulu\Component\Content\Structure\Page;
+use Sulu\Component\Content\Metadata;
+use Sulu\Component\Content\Metadata\Page;
 use Sulu\Component\Localization\Localization;
 use Sulu\Component\Webspace\Manager\WebspaceManagerInterface;
 use Sulu\Component\Webspace\Webspace;
