+++ src/Sulu/Component/Content/Compat/Serializer/PageBridgeHandler.php
@@ -17,7 +17,7 @@ use JMS\Serializer\Context;
 use JMS\Serializer\JsonDeserializationVisitor;
 use Sulu\Bundle\DocumentManagerBundle\Bridge\DocumentInspector;
 use Sulu\Component\Content\Compat\Structure\LegacyPropertyFactory;
-use Sulu\Component\Content\Structure\Factory\StructureFactory;
+use Sulu\Component\Content\Metadata\Factory\StructureFactory;
 use Sulu\Component\Content\Compat\Structure\PageBridge;
 use Sulu\Bundle\ContentBundle\Document\PageDocument;
 use Sulu\Component\DocumentManager\Document\UnknownDocument;
