+++ src/Sulu/Component/Content/Compat/Structure/LegacyPropertyFactory.php
@@ -5,12 +5,12 @@ namespace Sulu\Component\Content\Compat\Structure;
 use Sulu\Component\Content\Compat\Property as LegacyProperty;
 use Sulu\Component\Content\Compat\Section\SectionProperty;
 use Sulu\Component\Content\Compat\Block\BlockProperty;
-use Sulu\Component\Content\Structure\Property;
+use Sulu\Component\Content\Metadata\Property;
 use Sulu\Bundle\DocumentManagerBundle\Bridge\PropertyEncoder;
 use Sulu\Component\Content\Compat\Metadata;
-use Sulu\Component\Content\Structure\Section;
-use Sulu\Component\Content\Structure\Item;
-use Sulu\Component\Content\Structure\Block;
+use Sulu\Component\Content\Metadata\Section;
+use Sulu\Component\Content\Metadata\Item;
+use Sulu\Component\Content\Metadata\Block;
 use Sulu\Component\Content\Compat\Block\BlockPropertyType;
 use Sulu\Component\DocumentManager\NamespaceRegistry;
 use Sulu\Component\Content\Mapper\Translation\TranslatedProperty;
