+++ src/Sulu/Component/Content/Compat/Structure/StructureBridge.php
@@ -18,14 +18,14 @@ use Sulu\Component\Content\Document\ContentDocumentInterface;
 use Sulu\Component\Content\Document\LocalizationState;
 use Sulu\Component\Content\Document\RedirectType;
 use Sulu\Component\Content\Document\WorkflowStage;
-use Sulu\Component\Content\Structure\Factory\StructureFactory;
-use Sulu\Component\Content\Structure\Item;
-use Sulu\Component\Content\Structure\Property as NewProperty;
-use Sulu\Component\Content\Structure\Section;
-use Sulu\Component\Content\Structure\Structure;
+use Sulu\Component\Content\Metadata\Factory\StructureFactory;
+use Sulu\Component\Content\Metadata\Item;
+use Sulu\Component\Content\Metadata\Property as NewProperty;
+use Sulu\Component\Content\Metadata\Section;
+use Sulu\Component\Content\Metadata\Structure;
 use Sulu\Bundle\DocumentManagerBundle\Bridge\DocumentInspector;
 use Sulu\Component\Content\Document\Behavior\ContentBehavior;
-use Sulu\Component\Content\Structure\Block;
+use Sulu\Component\Content\Metadata\Block;
 use Sulu\Component\Content\Document\Behavior\ExtensionBehavior;
 use Sulu\Component\Content\Document\Behavior\ResourceSegmentBehavior;
 use Sulu\Component\Content\Document\Behavior\NavigationContextBehavior;
