+++ src/Sulu/Component/Content/Compat/StructureManager.php
@@ -12,9 +12,9 @@ namespace Sulu\Component\Content\Compat;
 
 use Symfony\Component\DependencyInjection\ContainerAware;
 use Sulu\Component\Content\Extension\ExtensionInterface;
-use Sulu\Component\Content\Structure\Factory\StructureFactory;
+use Sulu\Component\Content\Metadata\Factory\StructureFactory;
 use Sulu\Component\Content\Extension\ExtensionManager;
-use Sulu\Component\Content\Structure\Structure as NewStructure;
+use Sulu\Component\Content\Metadata\Structure as NewStructure;
 use Sulu\Component\Content\Compat\Structure\StructureBridge;
 use Sulu\Bundle\DocumentManagerBundle\Bridge\DocumentInspector;
 use Sulu\Component\Content\Compat\Structure\LegacyPropertyFactory;
