+++ src/Sulu/Component/Content/Exception/ExtensionNotFoundException.php
@@ -10,7 +10,7 @@
 
 namespace Sulu\Component\Content\Exception;
 
-use Sulu\Component\Content\StructureInterface;
+use Sulu\Component\Content\MetadataInterface;
 
 class ExtensionNotFoundException extends \Exception
 {
