+++ src/Sulu/Component/Content/Mapper/ContentMapper.php
@@ -48,9 +48,9 @@ use Sulu\Component\Content\Form\Exception\InvalidFormException;
 use Sulu\Component\Content\Mapper\Event\ContentNodeEvent;
 use Sulu\Component\Content\Mapper\Translation\MultipleTranslatedProperties;
 use Sulu\Component\Content\Mapper\Translation\TranslatedProperty;
-use Sulu\Component\Content\Structure\Page;
-use Sulu\Component\Content\Structure\Property;
-use Sulu\Component\Content\Structure\Structure;
+use Sulu\Component\Content\Metadata\Page;
+use Sulu\Component\Content\Metadata\Property;
+use Sulu\Component\Content\Metadata\Structure;
 use Sulu\Component\Content\Types\ResourceLocatorInterface;
 use Sulu\Component\Content\Types\Rlp\Strategy\RlpStrategyInterface;
 use Sulu\Component\DocumentManager\DocumentManager;
