+++ src/Sulu/Component/Content/Metadata/BlockMetadata.php
@@ -8,7 +8,7 @@
  * with this source code in the file LICENSE.
  */
 
-namespace Sulu\Component\Content\Structure;
+namespace Sulu\Component\Content\Metadata;
 
 /**
  * Block -- dynamic set of properties
