+++ src/Sulu/Component/Content/Metadata/Factory/StructureMetadataFactory.php
@@ -9,7 +9,7 @@
  * with this source code in the file LICENSE.
  */
  
-namespace Sulu\Component\Content\Structure\Factory;
+namespace Sulu\Component\Content\Metadata\Factory;
 
 use Symfony\Component\Config\ConfigCache;
 use Symfony\Component\Config\Resource\FileResource;
