+++ src/Sulu/Component/Content/Metadata/PageMetadata.php
@@ -8,13 +8,13 @@
  * with this source code in the file LICENSE.
  */
 
-namespace Sulu\Component\Content\Structure;
+namespace Sulu\Component\Content\Metadata;
 
 use JMS\Serializer\Annotation\Exclude;
 use JMS\Serializer\Annotation\Type;
 use Sulu\Component\Content\PageInterface;
-use Sulu\Component\Content\Structure;
-use Sulu\Component\Content\StructureInterface;
+use Sulu\Component\Content\Metadata;
+use Sulu\Component\Content\MetadataInterface;
 use Sulu\Component\Util\ArrayableInterface;
 
 /**
