+++ src/Sulu/Component/Content/Metadata/PropertyMetadata.php
@@ -9,7 +9,7 @@
  * with this source code in the file LICENSE.
  */
 
-namespace Sulu\Component\Content\Structure;
+namespace Sulu\Component\Content\Metadata;
 
 class PropertyMetadata extends Item
 {
