+++ src/Sulu/Component/Content/Metadata/SectionMetadata.php
@@ -9,7 +9,7 @@
  * with this source code in the file LICENSE.
  */
 
-namespace Sulu\Component\Content\Structure;
+namespace Sulu\Component\Content\Metadata;
 
 class SectionMetadata extends Item
 {
