+++ src/Sulu/Component/Content/StructureManager.php
@@ -11,9 +11,9 @@
 namespace Sulu\Component\Content;
 
 use Psr\Log\LoggerInterface;
-use Sulu\Component\Content\Structure\Page;
-use Sulu\Component\Content\Structure\Snippet;
-use Sulu\Component\Content\StructureExtension\StructureExtensionInterface;
+use Sulu\Component\Content\Metadata\Page;
+use Sulu\Component\Content\Metadata\Snippet;
+use Sulu\Component\Content\MetadataExtension\StructureExtensionInterface;
 use Sulu\Component\Content\Template\Dumper\PhpTemplateDumper;
 use Sulu\Component\Content\Template\Exception\InvalidXmlException;
 use Sulu\Component\Content\Template\Exception\TemplateNotFoundException;
@@ -87,9 +87,9 @@ class StructureManager extends ContainerAware implements StructureManagerInterfa
             'cache_dir' => null,
             'debug' => false,
             'page_cache_class_suffix' => 'PageCache',
-            'page_base_class' => 'Sulu\Component\Content\Structure\Page',
+            'page_base_class' => 'Sulu\Component\Content\Metadata\Page',
             'snippet_cache_class_suffix' => 'SnippetCache',
-            'snippet_base_class' => 'Sulu\Component\Content\Structure\Snippet',
+            'snippet_base_class' => 'Sulu\Component\Content\Metadata\Snippet',
         );
 
         // overwrite the default values with the given options
