+++ tests/Sulu/Component/Content/Compat/Structure/LegacyPropertyFactoryTest.php
@@ -5,14 +5,14 @@ namespace Sulu\Component\Content\Compat\Block\Structure;
 use Sulu\Component\Content\Compat\PropertyInterface;
 use Sulu\Component\Content\Compat\Structure\LegacyPropertyFactory;
 use Sulu\Bundle\DocumentManagerBundle\Bridge\PropertyEncoder;
-use Sulu\Component\Content\Structure\Property;
+use Sulu\Component\Content\Metadata\Property;
 use Sulu\Component\Content\Compat\PropertyInterface as LegacyPropertyInterface;
-use Sulu\Component\Content\Structure\Section;
+use Sulu\Component\Content\Metadata\Section;
 use Sulu\Component\Content\Compat\Metadata;
 use Sulu\Component\Content\Compat\Section\SectionPropertyInterface;
 use Sulu\Component\Content\Compat\Block\BlockProperty;
-use Sulu\Component\Content\Structure\Block;
-use Sulu\Component\Content\Structure\Component;
+use Sulu\Component\Content\Metadata\Block;
+use Sulu\Component\Content\Metadata\Component;
 use Sulu\Component\Content\Compat\Block\BlockPropertyInterface;
 use Sulu\Component\DocumentManager\NamespaceRegistry;
 use Sulu\Component\Content\Compat\PropertyParameter;
