+++ tests/Sulu/Component/Content/Compat/StructureManagerTest.php
@@ -2,10 +2,10 @@
 
 namespace Sulu\Component\Content\Compat;
 
-use Sulu\Component\Content\Structure\Factory\StructureFactory;
+use Sulu\Component\Content\Metadata\Factory\StructureFactory;
 use Sulu\Component\Content\Extension\ExtensionManager;
 use Sulu\Bundle\DocumentManagerBundle\Bridge\DocumentInspector;
-use Sulu\Component\Content\Structure\Structure;
+use Sulu\Component\Content\Metadata\Structure;
 use Sulu\Component\Content\Compat\Structure\StructureBridge;
 use Sulu\Component\Content\Extension\ExtensionInterface;
 use Sulu\Component\Content\Compat\Structure\LegacyPropertyFactory;
