+++ tests/Sulu/Component/Content/Document/Subscriber/ContentSubscriberTest.php
@@ -17,9 +17,9 @@ use Sulu\Component\Content\ContentTypeManagerInterface;
 use Sulu\Component\Content\Document\Behavior\ContentBehavior;
 use Sulu\Component\Content\Document\Property\PropertyContainer;
 use Sulu\Component\Content\Document\Subscriber\ContentSubscriber;
-use Sulu\Component\Content\Structure\Factory\StructureFactory;
-use Sulu\Component\Content\Structure\Property;
-use Sulu\Component\Content\Structure\Structure;
+use Sulu\Component\Content\Metadata\Factory\StructureFactory;
+use Sulu\Component\Content\Metadata\Property;
+use Sulu\Component\Content\Metadata\Structure;
 use Sulu\Component\DocumentManager\DocumentAccessor;
 use Sulu\Component\DocumentManager\Event\HydrateEvent;
 use Sulu\Component\DocumentManager\Event\PersistEvent;
