+++ tests/Sulu/Component/Content/Mapper/ContentMapperTest.php
@@ -23,8 +23,8 @@ use Sulu\Component\Content\Compat\Property;
 use Sulu\Component\Content\Compat\PropertyTag;
 use Sulu\Component\Content\Section\SectionProperty;
 use Sulu\Component\Content\Compat\Structure;
-use Sulu\Component\Content\StructureExtension\StructureExtension;
-use Sulu\Component\Content\StructureExtension\StructureExtensionInterface;
+use Sulu\Component\Content\MetadataExtension\StructureExtension;
+use Sulu\Component\Content\MetadataExtension\StructureExtensionInterface;
 use Sulu\Component\Content\Compat\StructureInterface;
 use Sulu\Component\Localization\Localization;
 use Sulu\Component\Webspace\Navigation;
